+++ constants/constants.go
@@ -24,3 +24,4 @@ package constants
 
 const DefaultDirMode = 0750
 const DefaultFileMode = 0750
+const WindowsExecutableExt = ".exe"